extract interface
reason: allow multiple implementations of the same interface to exist.
therefore, another impl (for calling external EMSP service) and bean
can exist with @Primary annotation which can take precedence
